Backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asarPy==1.0.1
bcrypt==4.3.0
click==8.2.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
motor==3.3.1
numpy==1.26.4
openai==1.12.0
passlib[bcrypt]==1.7.4
Pillow==10.0.0
prometheus-client==0.20.0
psutil==5.9.8
pydantic==2.11.7
pydantic[email]==2.11.7
pydantic_core==2.33.2
pymongo==4.14.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2024.1
redis==5.0.2
rq==1.16.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
bcrypt==4.3.0
email-validator==2.2.0 
httpx 
openai 
python-dotenv
validators==0.22.0
celery==5.3.6
flower==2.0.1

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.24.1
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
